% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayHist.R
\name{DayHist}
\alias{DayHist}
\title{Circular day histogram}
\usage{
DayHist(
  x,
  hist.zoom = 6,
  hist.col = "tomato",
  daynight.color = T,
  date = "03/20/2021",
  lon = -122.6396394,
  lat = 48.1639127,
  day.col = "lightgoldenrodyellow",
  night.col = "lightcyan"
)
}
\arguments{
\item{x}{A date-time object. Can be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or fts objects.  Essentially these are the same object types that can be used by the hour and minute functions from the lubridate package.}

\item{hist.zoom}{a zoom factor to appropriately scale histogram in plot area.}

\item{hist.col}{specifies the color the histogram bars will be}

\item{daynight.color}{a logical specifying if background coloration should be used to show daytime (between sunrise and sunset) and nighttime (between sunset and sunrise).}

\item{date}{Only used if daynight.color=TRUE. Specifies the date on which day-night background coloration time periods should be calculated.  Should follow the format "MM/DD/YYYY".}

\item{lon}{Only used if daynight.color=TRUE. Specifies the longitude of the location for which day-night background coloration should be calculated.  Should be in decimal format.}

\item{lat}{Only used if daynight.color=TRUE. Specifies the latitude of the location for which day-night background coloration should be calculated.  Should be in decimal format.}

\item{day.col}{Only used if daynight.color=TRUE. Specifies the color for which the daytime period of the background should be shaded.}

\item{night.col}{Only used if daynight.color=TRUE. Specifies the color for which the nighttime period of the background should be shaded.}
}
\value{
returns a histogram of the input times.
}
\description{
Will plot a series of times to a circular histogram over a 24-hour day.
}
